Correct measurement log for fast reading and :lipstick:
- responses_
QList<QString> --> QVector<QString> for faster operation.

To ensure that the measurement value is recorded in the correct
key, take the first response from the vector. However, the responses
should be cleared before calling any measurement command since
taking the first value may result in reading an unread response.

- getResponse()
Check if the list is empty. Otherwise, the program crushes.